+++ dsmc/octree.py
@@ -1,7 +1,6 @@
 import numpy as np
 import numpy.typing as npt
 from numba import njit
-import numba as nb
 
 fmin = np.finfo(float).min
 fmax = np.finfo(float).max
@@ -109,10 +108,6 @@ def _sort(permutations : npt.NDArray, box : npt.NDArray, positions : npt.NDArray
             Nnew += 1
             
     return new_permutations, Nnew
-
-@njit
-def get_V(box):
-    return (box[0][1] - box[0][0]) * (box[1][1] - box[1][0]) * (box[2][1] - box[2][0])
 
 @njit
 def _create_boxes(box):