+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2022-10-01
+### Added
+- inflow boundary condition
+- open boundary condition
+
 ## [0.6.0] - 2022-09-28
 ### Added
 - functionality for aspect ratio check in octree